+++ library/include/borealis/sidebar.hpp
@@ -32,7 +32,7 @@ class SidebarSeparator : public View
   public:
     SidebarSeparator();
 
-    void draw(NVGcontext* vg, int x, int y, unsigned width, unsigned height, Style* style, FrameContext* ctx) override;
+    void draw(NVGcontext* vg, float x, float y, float width, float height, Style* style, FrameContext* ctx) override;
 };
 
 class Sidebar;
@@ -50,7 +50,7 @@ class SidebarItem : public View
   public:
     SidebarItem(std::string label, Sidebar* sidebar);
 
-    void draw(NVGcontext* vg, int x, int y, unsigned width, unsigned height, Style* style, FrameContext* ctx) override;
+    void draw(NVGcontext* vg, float x, float y, float width, float height, Style* style, FrameContext* ctx) override;
 
     View* getDefaultFocus() override
     {
